One file per top-level class. Some types to vars. urls -> example.com
