+++ test/Microsoft.AspNet.Authentication.Test/OpenIdConnect/OpenIdConnectAuthenticationMiddlewareForTestingAuthenticate.cs
@@ -1,0 +1,45 @@
+﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using Microsoft.AspNet.Authentication.OpenIdConnect;
+using Microsoft.AspNet.Builder;
+using Microsoft.AspNet.DataProtection;
+using Microsoft.Framework.Logging;
+using Microsoft.Framework.OptionsModel;
+using Microsoft.Framework.WebEncoders;
+
+namespace Microsoft.AspNet.Authentication.Tests.OpenIdConnect
+{
+
+    /// <summary>
+    /// pass a <see cref="OpenIdConnectAuthenticationHandler"/> as the AuthenticationHandler
+    /// configured to handle certain messages.
+    /// </summary>
+    public class OpenIdConnectAuthenticationMiddlewareForTestingAuthenticate : OpenIdConnectAuthenticationMiddleware
+    {
+        OpenIdConnectAuthenticationHandler _handler;
+
+        public OpenIdConnectAuthenticationMiddlewareForTestingAuthenticate(
+            RequestDelegate next,
+            IOptions<OpenIdConnectAuthenticationOptions> options,
+            IDataProtectionProvider dataProtectionProvider,
+            ILoggerFactory loggerFactory,
+            IUrlEncoder encoder,
+            IOptions<ExternalAuthenticationOptions> externalOptions,
+            ConfigureOptions<OpenIdConnectAuthenticationOptions> configureOptions = null,
+            OpenIdConnectAuthenticationHandler handler = null
+            )
+        : base(next, options, dataProtectionProvider, loggerFactory, encoder, externalOptions, configureOptions)
+        {
+            _handler = handler;
+            var customFactory = loggerFactory as InMemoryLoggerFactory;
+            if (customFactory != null)
+                Logger = customFactory.Logger;
+        }
+
+        protected override AuthenticationHandler<OpenIdConnectAuthenticationOptions> CreateHandler()
+        {
+            return _handler ?? base.CreateHandler();
+        }
+    }
+}
